Post_App/requirements.txt
Django
django-cors-headers
pillow
PyJWT
djangorestframework
djangorestframework-simplejwt
django-environ